refactor: move scheduler types back to core
